chore: Move UserService files to UserService package and refactor imports
