--- backend/src/main/java/com/example/filmRanking/service/UserLoginService.java
@@ -1,4 +1,0 @@
-package com.example.filmRanking.service;
-
-public class UserLoginService {
-}
